refactor(xtask): organize workflow ownership (#1035)
## Intent

`xtask` has grown beyond docs generation, but its layout and ownership
still made the whole tree look docs-owned. This change separates the
workflow-specific code so repo automation remains centralized while docs
ownership is scoped to the docs generator only.

## Changes

- Move docs generation into `xtask/src/docs/`, including renderer, nav
updater, and snapshots.
- Extract shared source manifest discovery into `xtask/src/sources.rs`
for docs generation and truncation detection.
- Keep `xtask/src/main.rs` focused on CLI parsing and dispatch.
- Update CODEOWNERS so `@withcoral/repo-ops` owns all of `xtask`, with
`@withcoral/docs` added only for `xtask/src/docs/`.
- Narrow workflow-specific CI path filters for docs freshness,
performance checks, and skills sync.
- Update agent/docs guidance to describe the new ownership layout.

## Verification

- `cargo fmt --all -- --check`
- `cargo test -p xtask`
- `cargo clippy -p xtask --all-targets --locked -- -D warnings`
- `make docs-check`
- `git diff --check`
