+++ AGENTS.md
@@ -51,6 +51,12 @@
   installer supports.
 - Keep general repository automation in `xtask`; reserve `scripts/` for the
   bash Coral installer and installer-specific support.
+- Keep `xtask` organized by workflow: docs generation lives under
+  `xtask/src/docs/`, shared source-manifest discovery lives in
+  `xtask/src/sources.rs`, performance checks live in `xtask/src/perf.rs`, and
+  skill export lives in `xtask/src/skills.rs`. `@withcoral/repo-ops` owns the
+  whole tree; add narrower owners only for the workflow-specific paths they
+  actually maintain.
 - `make docs-check` intentionally skips the aggregate community source catalog.
   Any PR may leave that generated page stale so unrelated changes do not fail
   on aggregate community catalog drift; keep docs freshness strict for bundled
