+++ docs/AGENTS.md
@@ -22,8 +22,8 @@
   elsewhere in the repo, update the relevant docs in the same change
 - `docs/reference/bundled-sources.mdx` and
   `docs/reference/community-sources.mdx` are generated by
-  `make docs-generate`; update source manifests or the xtask renderer rather
-  than editing those pages by hand
+  `make docs-generate`; update source manifests or the renderer under
+  `xtask/src/docs/` rather than editing those pages by hand
 - `make docs-check` intentionally skips
   `docs/reference/community-sources.mdx`; any PR may leave that generated page
   stale so unrelated changes do not fail on aggregate community catalog drift.
